move Post->PostForm, add Post method
Post opperations to the chef-server that require data need an io object
The current chef.Post implements the behavior of PostForm from standard
library so I renamed it to chef.PostForm, and added chef.Post. chef.Post
takes an io object and sends it to the server with the signed headders.